Bugfixes, clarifications.
The bidirectional type system apparently needed a last ctxApply step so
as to show the correct types; as it stood λx. x was giving the type
∃a -> ∃b rather than ∃a -> ∃a, which seems incorrect to the end user if
they can't see the context.

Additionally, some simplifications were made to the code, and I
clarified some precedence using parenthesis. Previously I wrote "forall
a. a <: 1" which could be misinterpreted as "forall a. (a <: 1)" rather
than the intended "(forall a. a) <: 1".